+++ buildDocs.ps1
@@ -1,2 +1,2 @@
 dotnet tool update -g docfx
-docfx .\docs\docfx.json --serve
+docfx .\docs\docfx.json --serve --port 9000