+++ docs/docs/bots/code/WatcherBot/Step4/ValidateBotOptions.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Extensions.Options;
+
+namespace WatcherBot
+{
+    [OptionsValidator]
+    public partial class ValidateBotOptions : IValidateOptions<BotOptions>
+    {
+    }
+}