+++ docs/docs/bots/code/WatcherBot/Step5/BotOptions.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace WatcherBot
+{
+    public sealed class BotOptions
+    {
+        public const string ConfigurationSectionName = "Bot";
+
+        [Required]
+        public required string[] WatchWords { get; set; }
+    }
+}